refactor to get clean code
