+++ crates/pipe-exec-layer-ext-v2/execute/src/onchain_config/epoch.rs
@@ -1,6 +1,6 @@
 use super::base::{ConfigFetcher, OnchainConfigFetcher};
 use super::{EPOCH_MANAGER_ADDR, SYSTEM_CALLER};
-use alloy_primitives::{Address, Bytes, U256};
+use alloy_primitives::{Address, Bytes};
 use alloy_sol_macro::sol;
 use alloy_sol_types::SolCall;
 use reth_rpc_eth_api::helpers::EthCall;
